Refactor large Go files into smaller units
